requirements.txt
```
requirements
streamlit
opencv-python
pillow
mediapipe
numpy
